# server/openapi.yaml
...
openapi: 3.1.0
info:
  title: Fake Stack Overflow API
  version: 1.0.0
  description: OpenAPI 3.1 specification for the Fake Stack Overflow backend server.
servers:
  - url: /api
    description: The API server endpoint
paths:
...
  /chat/createChat:
    post:
      tags:
        - Chats
      summary: Create a new chat conversation.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateChatBody'
      responses:
        '200':
          description: Chat successfully created. Returns the populated chat object.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PopulatedChat'
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/ErrorResponse'
  /chat/{chatId}/addMessage:
    post:
      tags:
        - Chats
      summary: Add a new message to an existing chat.
      parameters:
        - in: path
          name: chatId
          schema:
            type: string
            format: object-id
          required: true
          description: The unique ID of the chat.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MessageBody'
      responses:
        '200':
          description: Message successfully added. Returns the updated populated chat object.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PopulatedChat'
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/ErrorResponse'
  /chat/{chatId}:
    get:
      tags:
        - Chats
      summary: Retrieve a single chat by its ID.
      parameters:
        - in: path
          name: chatId
          schema:
            type: string
            format: object-id
          required: true
          description: The unique ID of the chat.
      responses:
        '200':
          description: Successfully retrieved chat details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PopulatedChat'
        '500':
          $ref: '#/components/responses/ErrorResponse'
  /chat/{chatId}/addParticipant:
    post:
      tags:
        - Chats
      summary: Add a new participant to an existing chat.
      parameters:
        - in: path
          name: chatId
          schema:
            type: string
            format: object-id
          required: true
          description: The unique ID of the chat.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddParticipantBody'
      responses:
        '200':
          description: Participant successfully added. Returns the updated database chat object.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PopulatedChat'
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/ErrorResponse'
  /chat/getChatsByUser/{username}:
    get:
      tags:
        - Chats
      summary: Retrieve all chats involving a specific user.
      parameters:
        - in: path
          name: username
          schema:
            type: string
          required: true
          description: The username whose chats are being retrieved.
      responses:
        '200':
          description: Successfully retrieved list of chats.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PopulatedChat'
        '500':
          $ref: '#/components/responses/ErrorResponse'

  # --------------------
  # GAME CONTROLLER
  # --------------------
...
components:
  schemas:
...
    MessageBody:
      type: object
      required:
        - msg
        - msgFrom
        - msgDateTime
      properties:
        msg:
          type: string
          minLength: 1
        msgFrom:
          type: string
          minLength: 1
        msgDateTime:
          type: string
          format: date-time
...
    CreateChatBody:
      type: object
      required:
        - participants
        - messages
      properties:
        participants:
          type: array
          minItems: 1
          items:
            type: string
        messages:
          type: array
          items:
            $ref: '#/components/schemas/MessageBody'
    AddParticipantBody:
      type: object
      required:
        - username
      properties:
        username:
          type: string
          description: The username of the new participant.
    PopulatedChat:
      type: object
      description: A fully populated chat object with embedded messages.

    # --------------------
    # GAME SCHEMAS
    # --------------------
...
  responses:
...
    ErrorResponse:
      description: Internal Server Error
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                description: The error message returned by the server.
    ValidationError:
      description: Bad Request / Validation Error
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                description: The error message from express-openapi-validator.
              errors:
                type: array
                description: Detailed list of validation errors.